Global tests moved to neutral package
